Split classes in file.h and value_stores.* to separate files
